Parametrise tests with invalid permissions
