+++ berth_reservations/tests/factories.py
@@ -1,0 +1,12 @@
+import factory
+from django.contrib.auth import get_user_model
+
+
+class UserFactory(factory.django.DjangoModelFactory):
+    first_name = factory.Faker("first_name")
+    last_name = factory.Faker("last_name")
+    username = factory.Faker("user_name")
+    email = factory.Faker("email")
+
+    class Meta:
+        model = get_user_model()
