+++ resources/tests/test_resources_mutations.py
@@ -1,5 +1,6 @@
 import uuid
 
+import pytest
 from graphql_relay import to_global_id
 
 from berth_reservations.tests.utils import (
@@ -51,7 +52,8 @@ def test_create_berth(pier, berth_type, superuser):
     assert executed["data"]["createBerth"]["berth"]["number"] == "9999"
 
 
-def test_create_berth_no_user(pier, berth_type):
+@pytest.mark.parametrize("user", ["none", "base", "staff"], indirect=True)
+def test_create_berth_not_enough_permissions(user, pier, berth_type):
     variables = {
         "number": "9999",
         "comment": "foobar",
@@ -72,28 +74,7 @@ def test_create_berth_no_user(pier, berth_type):
     assert_not_enough_permissions(executed)
 
 
-def test_create_berth_staff_user(pier, berth_type, staff_user):
-    variables = {
-        "number": "9999",
-        "comment": "foobar",
-        "pierId": to_global_id("PierNode", str(pier.id)),
-        "berthTypeId": to_global_id("BerthTypeNode", str(berth_type.id)),
-    }
-
-    assert Berth.objects.count() == 0
-
-    executed = client.execute(
-        query=CREATE_BERTH_MUTATION,
-        variables=variables,
-        graphql_url=GRAPHQL_URL,
-        user=None,
-    )
-
-    assert Berth.objects.count() == 0
-    assert_not_enough_permissions(executed)
-
-
-def test_create_berth_no_number(pier, berth_type, staff_user):
+def test_create_berth_no_number(pier, berth_type, superuser):
     variables = {
         "pierId": to_global_id("PierNode", str(pier.id)),
         "berthTypeId": to_global_id("BerthTypeNode", str(berth_type.id)),
@@ -105,7 +86,7 @@ def test_create_berth_no_number(pier, berth_type, staff_user):
         query=CREATE_BERTH_MUTATION,
         variables=variables,
         graphql_url=GRAPHQL_URL,
-        user=None,
+        user=superuser,
     )
 
     assert Berth.objects.count() == 0
@@ -121,7 +102,7 @@ DELETE_BERTH_MUTATION = """
 """
 
 
-def test_delete_berth_superuser(superuser, berth):
+def test_delete_berth(superuser, berth):
     variables = {
         "id": to_global_id("BerthNode", str(berth.id)),
     }
@@ -138,7 +119,8 @@ def test_delete_berth_superuser(superuser, berth):
     assert Berth.objects.count() == 0
 
 
-def test_delete_berth_staff_user(staff_user, berth):
+@pytest.mark.parametrize("user", ["none", "base", "staff"], indirect=True)
+def test_delete_berth_not_enough_permissions(user, berth):
     variables = {
         "id": to_global_id("BerthNode", str(berth.id)),
     }
@@ -149,25 +131,7 @@ def test_delete_berth_staff_user(staff_user, berth):
         query=DELETE_BERTH_MUTATION,
         variables=variables,
         graphql_url="/graphql_v2/",
-        user=staff_user,
-    )
-
-    assert Berth.objects.count() == 1
-    assert_not_enough_permissions(executed)
-
-
-def test_delete_berth_no_user(berth):
-    variables = {
-        "id": to_global_id("BerthNode", str(berth.id)),
-    }
-
-    assert Berth.objects.count() == 1
-
-    executed = client.execute(
-        query=DELETE_BERTH_MUTATION,
-        variables=variables,
-        graphql_url="/graphql_v2/",
-        user=None,
+        user=user,
     )
 
     assert Berth.objects.count() == 1
@@ -260,3 +224,27 @@ def test_update_berth_no_id(berth, pier, berth_type, superuser):
 
     assert Berth.objects.count() == 1
     assert_field_missing("id", executed)
+
+
+@pytest.mark.parametrize("user", ["none", "base", "staff"], indirect=True)
+def test_update_berth_not_enough_permissions(berth, pier, berth_type, user):
+    pier_id = to_global_id("PierNode", str(pier.id))
+    berth_type_id = to_global_id("BerthTypeNode", str(berth_type.id))
+
+    variables = {
+        "number": "666",
+        "comment": "foobar",
+        "pierId": pier_id,
+        "berthTypeId": berth_type_id,
+    }
+    assert Berth.objects.count() == 1
+
+    executed = client.execute(
+        query=CREATE_BERTH_MUTATION,
+        variables=variables,
+        graphql_url=GRAPHQL_URL,
+        user=user,
+    )
+
+    assert Berth.objects.count() == 1
+    assert_not_enough_permissions(executed)
